Move utils and constants
